.history/app/backend/requirements_20240920104153.txt
fastapi
uvicorn
opencv-python
numpy
RPi.GPIO
pydantic